DO NOT MERGE INTO MASTER
Update the Docker files to work with a local setup where the edgex-core
services are running on my local machine. This required updates to the
URLs used by the docker containers so that they can loop-back to my
Linux host.

Create a test script which will build and run from edgex from source,
start the necessary containers, and run the blackbox tests.

NOTE the black-box tests for device-virtual and rules-engine have been
disabled as there is issues with running them with this setup. This is
due to the networking being hard-coded in the Docker images and can't be
updated to use the loop-back address to communicate with core-services
running on my host Linux machine.